Fixes #142: Move indexing logic entirely to the server-side
Indexing is now triggered as part of the onInitialize() request.

Any code on the client related to indexing (except registered VS Code
commands) is moved to the server. The user commands are just proxies -
they make a request to the server.

All the code related to building, querying and maintaining the index is
moved to the index.ts file.

Wherever possible, standard message request from the language server
protocol are used instead of custom messages. For example
onDidChangeContent() and onDidChangeWatchedFiles() instead of the custom
"buildObjectTreeForDocument", "saveTreeCache", etc.